backend/pip_requirements.txt
appdirs
appnope
asn1crypto
attrs
Automat
backcall
bleach
certifi
cffi
chardet
clyent
constantly
cryptography
cycler
Cython
decorator
entrypoints
html5lib
hyperlink
idna
incremental
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
kiwisolver
MarkupSafe
matplotlib
mistune
#mkl-fft
#mkl-random
#nb-anacondacloud
#nb-conda
#nb-conda-kernels
nbconvert
#nbformat
#nbpresent
#notebook
numpy
olefile
pandas
pandocfilters
parso
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
ptyprocess
pyasn1
pyasn1-modules
pycparser
Pygments
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytz
PyYAML
pyzmq
requests
scikit-learn
scipy
Send2Trash
service-identity
simplegeneric
six
terminado
testpath
torch
torchvision
tornado
traitlets
Twisted
urllib3
wcwidth
webencodings
widgetsnbextension
zope.interface
